requirements.txt
tk
pillow
sqlite3
python-dotenv
smtplib
email